divide three configs for webpack
