+++ package.json
@@ -1,85 +1,87 @@
 {
-  "name": "steamer-react",
-  "version": "0.0.1",
-  "description": "react-redux-webpack-boilerplate",
-  "main": "index.js",
-  "scripts": {
-    "dev": "gulp&&node ./server.js",
-    "pub-mac": "gulp sprites&&export NODE_ENV=production&&webpack",
-    "pub": "gulp sprites&&set NODE_ENV=production&&webpack"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/lcxfs1991/steamer-react.git"
-  },
-  "keywords": [
-    "react",
-    "redux",
-    "webpack",
-    "boilerplate"
-  ],
-  "author": "lcxfs1991",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/lcxfs1991/steamer-react/issues"
-  },
-  "homepage": "https://github.com/lcxfs1991/steamer-react#readme",
-  "dependencies": {
-    "react": "^15.1.0",
-    "react-dom": "^15.1.0",
-    "react-redux": "^4.0.6",
-    "react-router": "^2.4.1",
-    "react-router-redux": "^4.0.5",
-    "react-immutable-render-mixin": "^0.9.5",
-    "redux": "^3.2.0",
-    "classnames": "^2.2.3",
-    "redux-immutable": "^3.0.6",
-    "redux-thunk": "^1.0.3",
-    "history": "^3.0.0",
-    "immutable": "^3.7.6",
-    "lodash.merge": "^4.4.0"
-  },
-  "devDependencies": {
-    "babel-core": "^6.1.4",
-    "babel-loader": "^6.1.0",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-preset-es2015": "^6.1.4",
-    "babel-preset-es2015-loose": "^7.0.0",
-    "babel-preset-react": "^6.1.4",
-    "clean-webpack-plugin": "^0.1.4",
-    "copy-webpack-plugin": "^1.1.1",
-    "css-loader": "^0.22.0",
-    "exports-loader": "^0.6.2",
-    "expose-loader": "^0.7.1",
-    "express": "^4.13.4",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.8.4",
-    "gulp": "^3.9.0",
-    "gulp-replace": "^0.5.4",
-    "gulp-zip": "^3.0.2",
-    "gulp.spritesmith-multi": "^3.0.0",
-    "html-loader": "^0.3.0",
-    "html-res-webpack-plugin": "^0.0.6",
-    "http-proxy": "^1.12.0",
-    "image-webpack-loader": "^1.8.0",
-    "imports-loader": "^0.6.3",
-    "merge-stream": "^1.0.0",
-    "node-sass": "^3.4.1",
-    "proxy-middleware": "^0.15.0",
-    "react-addons-perf": "^15.1.0",
-    "react-hot-loader": "^1.3.0",
-    "redux-devtools": "^3.0.1",
-    "redux-devtools-dock-monitor": "^1.0.1",
-    "redux-devtools-log-monitor": "^1.0.2",
-    "run-sequence": "^1.1.5",
-    "sass-loader": "^3.1.1",
-    "style-loader": "^0.13.0",
-    "url-loader": "^0.5.6",
-    "webpack": "^1.13.0",
-    "webpack-dev-middleware": "^1.6.1",
-    "webpack-dev-server": "^1.14.1",
-    "webpack-hot-middleware": "^2.10.0",
-    "webpack-md5-hash": "^0.0.5",
-    "yargs": "^4.7.0"
-  }
+    "name": "steamer-react",
+    "version": "0.0.1",
+    "description": "react-redux-webpack-boilerplate",
+    "main": "index.js",
+    "scripts": {
+        "dev": "gulp&&node ./webpack.server.js",
+        "pub-mac": "gulp sprites&&export NODE_ENV=production&&webpack",
+        "pub": "gulp sprites&&set NODE_ENV=production&&webpack"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/lcxfs1991/steamer-react.git"
+    },
+    "keywords": [
+        "react",
+        "redux",
+        "webpack",
+        "boilerplate"
+    ],
+    "author": "lcxfs1991",
+    "license": "MIT",
+    "bugs": {
+        "url": "https://github.com/lcxfs1991/steamer-react/issues"
+    },
+    "homepage": "https://github.com/lcxfs1991/steamer-react#readme",
+    "dependencies": {},
+    "devDependencies": {},
+    "dependenciesBk": {
+        "react": "^15.1.0",
+        "react-dom": "^15.1.0",
+        "react-redux": "^4.0.6",
+        "react-router": "^2.4.1",
+        "react-router-redux": "^4.0.5",
+        "react-immutable-render-mixin": "^0.9.5",
+        "redux": "^3.2.0",
+        "classnames": "^2.2.3",
+        "redux-immutable": "^3.0.6",
+        "redux-thunk": "^1.0.3",
+        "history": "^3.0.0",
+        "immutable": "^3.7.6",
+        "lodash.merge": "^4.4.0"
+    },
+    "devDependenciesBk": {
+        "babel-core": "^6.1.4",
+        "babel-loader": "^6.1.0",
+        "babel-plugin-transform-decorators-legacy": "^1.3.4",
+        "babel-preset-es2015": "^6.1.4",
+        "babel-preset-es2015-loose": "^7.0.0",
+        "babel-preset-react": "^6.1.4",
+        "clean-webpack-plugin": "^0.1.4",
+        "copy-webpack-plugin": "^1.1.1",
+        "css-loader": "^0.22.0",
+        "exports-loader": "^0.6.2",
+        "expose-loader": "^0.7.1",
+        "express": "^4.13.4",
+        "extract-text-webpack-plugin": "^1.0.1",
+        "file-loader": "^0.8.4",
+        "gulp": "^3.9.0",
+        "gulp-replace": "^0.5.4",
+        "gulp-zip": "^3.0.2",
+        "gulp.spritesmith-multi": "^3.0.0",
+        "html-loader": "^0.3.0",
+        "html-res-webpack-plugin": "^0.0.6",
+        "http-proxy": "^1.12.0",
+        "image-webpack-loader": "^1.8.0",
+        "imports-loader": "^0.6.3",
+        "merge-stream": "^1.0.0",
+        "node-sass": "^3.4.1",
+        "proxy-middleware": "^0.15.0",
+        "react-addons-perf": "^15.1.0",
+        "react-hot-loader": "^1.3.0",
+        "redux-devtools": "^3.0.1",
+        "redux-devtools-dock-monitor": "^1.0.1",
+        "redux-devtools-log-monitor": "^1.0.2",
+        "run-sequence": "^1.1.5",
+        "sass-loader": "^3.1.1",
+        "style-loader": "^0.13.0",
+        "url-loader": "^0.5.6",
+        "webpack": "^1.13.0",
+        "webpack-dev-middleware": "^1.6.1",
+        "webpack-dev-server": "^1.14.1",
+        "webpack-hot-middleware": "^2.10.0",
+        "webpack-md5-hash": "^0.0.5",
+        "yargs": "^4.7.0"
+    }
 }
